requirements.txt
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
decorator==4.4.2
docutils==0.22.4
ez_setup==0.9
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
id==1.6.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
keyring==25.7.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.6.1
nh3==0.3.2
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pillow==11.3.0
proglog==0.1.12
progress==1.6.1
psutil==7.2.2
Pygments==2.19.2
python-dotenv==1.2.1
pywin32-ctypes==0.2.3
PyYAML==6.0.3
readme_renderer==44.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.3.2
safetensors==0.7.0
scipy==1.17.0
setuptools==80.10.2
shellingham==1.5.4
static_ffmpeg==3.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.0.0
twine==6.2.0
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
yt-dlp==2026.2.4
